refactor controllers into to services + delete api.py
