+++ dinghy_ping/controllers/PingController.py
@@ -1,24 +1,13 @@
-import requests
-import logging
-from urllib.parse import urlparse
-import json
-from prometheus_client import Counter, Summary
-import dinghy_ping.models.data as data
-
+from dinghy_ping.services.PingService import PingService
+from dinghy_ping.services.RequestService import RequestService
 api = None
-
-# Prometheus metrics
-COMPLETED_REQUEST_COUNTER = Counter('dingy_pings_completed', 'Count of completed dinghy ping requests')
-FAILED_REQUEST_COUNTER = Counter('dingy_pings_failed', 'Count of failed dinghy ping requests')
-REQUEST_TIME = Summary('dinghy_request_processing_seconds', 'Time spent processing request')
-
 
 class PingController:
 
   def __init__(self, api, redis_host):
-    self.redis_host = redis_host
+    self.request_service = RequestService(redis_host)
+    self.ping_service = PingService(redis_host)
     global api
-    self.api = api
     api = api
    
   @api.route("/")
@@ -26,16 +15,8 @@ class PingController:
       """Index route to Dinghy-ping input html form"""
       resp.content = api.template(
           'index.html',
-          get_all_pinged_urls=self._get_all_pinged_urls()
+          get_all_pinged_urls=self.ping_service._get_all_pinged_urls()
       ) 
-
-
-  def _get_all_pinged_urls(self):
-    """Get pinged URLs from Dinghy-ping data module"""
-    p = data.DinghyData(self.redis_host)
-
-    return p.get_all_pinged_urls() 
-
 
   @api.route("/ping/{protocol}/{domain}")
   def domain_response_html(self, req, resp, *, protocol, domain):
@@ -46,7 +27,7 @@ class PingController:
 
       headers = {}
       domain_response_code, domain_response_text, domain_response_time_ms, domain_response_headers = (
-          self._process_request(protocol, domain, req.params, headers)
+          self.request_service._process_request(protocol, domain, req.params, headers)
       )
 
       resp.content = api.template(
@@ -57,48 +38,6 @@ class PingController:
               domain_response_headers=domain_response_headers,
               domain_response_time_ms=domain_response_time_ms
       )
-
-
-  @REQUEST_TIME.time()
-  def _process_request(self, protocol, domain, params, headers):
-      """
-      Internal method to run request process, takes protocol and domain for input
-      """
-
-      if protocol == "":
-          protocol = "https"
-
-      domain_response_code = ""
-      domain_response_text = ""
-      domain_response_time_ms = ""
-      domain_response_headers = {}
-
-      try:
-          r = requests.get(f'{protocol}://{domain}', params=params, timeout=5, headers=headers)
-          COMPLETED_REQUEST_COUNTER.inc()
-      except requests.exceptions.Timeout as err:
-          domain_response_text = f'Timeout: {err}'
-          FAILED_REQUEST_COUNTER.inc()
-          return domain_response_code, domain_response_text, domain_response_time_ms, domain_response_headers
-      except requests.exceptions.TooManyRedirects as err:
-          domain_response_text = f'TooManyRedirects: {err}'
-          FAILED_REQUEST_COUNTER.inc()
-          return domain_response_code, domain_response_text, domain_response_time_ms, domain_response_headers
-      except requests.exceptions.RequestException as err:
-          domain_response_text = f'RequestException: {err}'
-          FAILED_REQUEST_COUNTER.inc()
-          return domain_response_code, domain_response_text, domain_response_time_ms, domain_response_headers
-
-      domain_response_code = r.status_code
-      domain_response_text = r.text
-      domain_response_headers = dict(r.headers)
-      domain_response_time_ms = r.elapsed.microseconds / 1000
-      print(domain_response_headers)
-
-      d = data.DinghyData(self.redis_host, domain_response_code, domain_response_time_ms, r.url)
-      d.save_ping()
-
-      return domain_response_code, domain_response_text, domain_response_time_ms, domain_response_headers
     
   @api.route("/ping/domains")
   async def ping_multiple_domains(self, req, resp):
@@ -139,22 +78,3 @@ class PingController:
       }
       """
       # results = []
-
-# def build_domain_results(protocol, request_domain, results, headers):
-#     domain_response_code, domain_response_text, domain_response_time_ms, domain_response_headers = _process_request(protocol, request_domain, req.params, headers)
-#     results.append({
-#         "protocol": protocol,
-#         "domain": request_domain,
-#         "domain_response_code": domain_response_code,
-#         "domain_response_headers": domain_response_headers,
-#         "domain_response_time_ms": domain_response_time_ms
-#     })
-
-# def gather_results(data):
-#     for domain in data['domains']:
-#         protocol = domain['protocol']
-#         request_domain = domain['domain']
-#         headers = domain['headers']
-#         build_domain_results(protocol, request_domain, results, headers)
-
-# resp.media = {"domains_response_results": results, "wait": gather_results(await req.media())}
